Add missing value sets to value set tabs
