+++ print_elements.js
@@ -1,5 +1,38 @@
 const { Identifier, IdentifiableValue, ChoiceValue, TypeConstraint, CardConstraint, IncludesTypeConstraint, ConstraintsFilter, RefValue, Cardinality } = require('shr-models');
 
+const urlsToNames = {
+  'https://sdt.cap.org': 'College of American Pathologists',
+  'http://www.dsm5.org/': 'DSM-5',
+  'https://evs.nci.nih.gov/ftp1/CDISC/SDTM/': 'CDISC SDTM Controlled Terminology',
+  'http://www.genenames.org': 'Hugo Gene Nomenclature Committee',
+  'http://hl7.org/fhir/quantity-comparator': 'FHIR Quantity Comparator',
+  'http://hl7.org/fhir/sid/cvx': 'CVX',
+  'http://hl7.org/fhir/allergy-verification-status': 'FHIR Allergy Intolerance Verification Status',
+  'http://hl7.org/fhir/observation-status': 'FHIR Observation Status',
+  'http://hl7.org/fhir/ValueSet/allergy-intolerance-category': 'FHIR Allergy Intolerance Category',
+  'http://hl7.org/fhir/ValueSet/allergy-intolerance-type': 'FHIR Allergy Intolerance Category',
+  'http://hl7.org/fhir/observation-category': 'FHIR Observation Category',
+  'http://hl7.org/fhir/v3/ActReason': 'HL7 V3 ActReason',
+  'http://hl7.org/fhir/v3/ObservationInterpretation': 'HL7 V3 Observation Interpretation',
+  'http://hl7.org/fhir/v2/0487': 'HL7 V2 Specimen Type',
+  'http://loinc.org': 'LOINC',
+  'http://www.meddra.org': 'MedDRA',
+  'http://www.nationsonline.org/oneworld/country_code_list': 'CC',
+  'https://www.ncbi.nlm.nih.gov/refseq': 'NCBI Reference Sequence Database',
+  'http://ncimeta.nci.nih.gov': 'NCI Metathesaurus',
+  'http://uts.nlm.nih.gov/metathesaurus': 'NCI Metatheasurus',
+  'https://ncit.nci.nih.gov': 'NCI Thesaurus',
+  'https://ncit.nci.nih.gov/ncitbrowser/ConceptReport.jsp?dictionary=NCI_Thesaurus': 'NCI Thesaurus',
+  'http://www.nlm.nih.gov/research/umls/rxnorm': 'RxNorm',
+  'http://snomed.info/sct': 'SNOMED CT',
+  'http://hl7.org/fhir/sid/icd-10-cm': 'ICD-10-CM',
+  'http://unitsofmeasure.org': 'UCUM',
+  'http://codes.iarc.fr/topography': 'ICD-O-3 Topology Codes',
+  'urn:iso:std:iso:4217': 'CURRENCY',
+  'urn:tbd:': 'TBD',
+  'urn:tbd': 'TBD'
+};
+
 // These will insert a space in between:
 // - not a capital letter -- a capital letter
 // - a capital letter -- a capital letter follow by not a capital letter
@@ -42,7 +75,7 @@ function getDataType(de) {
   return type;
 }
 
-function getBinding(de, path, specs, projectURL) {
+function getVsInfo(de, path, specs, projectURL) {
   const value = findValueByPath(specs, path, de);
   let constraint;
   const vsConstraints = value.constraintsFilter.valueSet.constraints;
@@ -66,7 +99,8 @@ function getBinding(de, path, specs, projectURL) {
   }
   if (constraint) {
     const url = constraint.valueSet.startsWith(projectURL) ? constraint.valueSet.split('/')[constraint.valueSet.split('/').length-1] : constraint.valueSet;
-    return { url, strength: constraint.bindingStrength };
+    const vs = specs.valueSets.findByURL(constraint.valueSet);
+    return { url, strength: constraint.bindingStrength, vs };
   }
 }
 
@@ -292,7 +326,7 @@ function aggregateCardinality(...card) {
   return new Cardinality(min, max);
 }
 
-function fillLines(dataElementLines, profileLines, de, specs, config) {
+function fillElementLines(dataElementLines, profileLines, vsMap, de, specs, config) {
   let isInProfileList = false;
   const valueAndFields = [de.value, ...de.fields];
   const profileName = getHumanReadableProfileName(de.identifier.name);
@@ -321,11 +355,14 @@ function fillLines(dataElementLines, profileLines, de, specs, config) {
         const card = getAggregateEffectiveCardinality(de.identifier, rule.path, specs);
         const cardInfo = getCardinalityInfo(card);
         const dataType = getDataType(endOfPathElement);
-        const binding = getBinding(de, rule.path, specs, config.projectURL);
-        const url = binding ? binding.url : '';
-        const strength = binding ? binding.strength.toLowerCase() : '';
+        const vsInfo = getVsInfo(de, rule.path, specs, config.projectURL);
+        const url = vsInfo ? vsInfo.url : '';
+        const strength = vsInfo ? vsInfo.strength.toLowerCase() : '';
         const unit = getUnit(de, rule.path, specs, config.projectURL);
         dataElementLines.push([profileName, pathName, description, cardInfo.requirement, cardInfo.multiplicity, dataType, url, strength, unit]);
+        if (vsInfo) {
+          vsMap.set(vsInfo.url, vsInfo.vs);
+        }
       }
     }
   }
@@ -335,12 +372,86 @@ function fillLines(dataElementLines, profileLines, de, specs, config) {
   }
 }
 
-module.exports = function printElements(specs, config) {
-  const dataElementLines = [['Profile Name', 'Data Element Name', 'Description', 'Required in Profile?', 'Occurrences Allowed', 'Data Type', 'Value Set', 'Value Set Binding', 'Units']];
-  const profileLines = [['Profile Name', 'Profile Description']];
-  for (const de of specs.dataElements.all) {
-    fillLines(dataElementLines, profileLines, de, specs, config);
+function fillValueSetLines(vs, valueSetLines, valueSetDetailsLines) {
+  for (const rule of vs.rulesFilter.includesCode.rules) {
+    valueSetDetailsLines.push([
+      vs.identifier.name,
+      urlsToNames[rule.code.system] ? urlsToNames[rule.code.system] : rule.code.system,
+      `${rule.code.code}`,
+      `${rule.code.display}`
+    ]);
+  }
+  for (const rule of vs.rulesFilter.includesDescendents.rules) {
+    valueSetDetailsLines.push([
+      vs.identifier.name,
+      urlsToNames[rule.code.system] ? urlsToNames[rule.code.system] : rule.code.system,
+      `includes codes descending from ${rule.code.code}`,
+      `${rule.code.display}`
+    ]);
+  }
+  for (const rule of vs.rulesFilter.includesFromCode.rules) {
+    valueSetDetailsLines.push([
+      vs.identifier.name,
+      urlsToNames[rule.code.system] ? urlsToNames[rule.code.system] : rule.code.system,
+      `includes codes from code ${rule.code.code}`,
+      `${rule.code.display}`
+    ]);
+  }
+  for (const rule of vs.rulesFilter.includesFromCodeSystem.rules) {
+    valueSetDetailsLines.push([
+      vs.identifier.name,
+      urlsToNames[rule.system] ? urlsToNames[rule.system] : rule.system,
+      `includes codes from code system ${urlsToNames[rule.system] ? urlsToNames[rule.system] : rule.system}`,
+      ''
+    ]);
+  }
+  for (const rule of vs.rulesFilter.excludesDescendents.rules) {
+    valueSetDetailsLines.push([
+      vs.identifier.name,
+      urlsToNames[rule.code.system] ? urlsToNames[rule.code.system] : rule.code.system,
+      `excludes codes descending from ${rule.code.code}`,
+      `${rule.code.display}`
+    ]);
   }
 
-  return { profileLines, dataElementLines };
+  let codeSystems = new Set();
+  for (const rule of vs.rulesFilter.includesCode.rules) {
+    codeSystems.add(urlsToNames[rule.code.system]);
+  }
+  for (const rule of vs.rulesFilter.includesDescendents.rules) {
+    codeSystems.add(urlsToNames[rule.code.system]);
+  }
+  for (const rule of vs.rulesFilter.includesFromCode.rules) {
+    codeSystems.add(urlsToNames[rule.code.system]);
+  }
+  for (const rule of vs.rulesFilter.includesFromCodeSystem.rules) {
+    codeSystems.add(urlsToNames[rule.system]);
+  }
+  for (const rule of vs.rulesFilter.excludesDescendents.rules) {
+    codeSystems.add(urlsToNames[rule.code.system]);
+  }
+  codeSystems.delete(null);
+  codeSystems.delete(undefined);
+  valueSetLines.push([
+    vs.identifier.name,
+    `${vs.description ? vs.description : ''}`,
+    `${Array.from(codeSystems).join(', ')}`
+  ]);
+}
+
+module.exports = function printElements(specs, config) {
+  const vsMap = new Map();
+  const dataElementLines = [['Profile Name', 'Data Element Name', 'Description', 'Required in Profile?', 'Occurrences Allowed', 'Data Type', 'Value Set', 'Value Set Binding', 'Units']];
+  const profileLines = [['Profile Name', 'Profile Description']];
+  const valueSetLines = [['Value Set', 'Description', 'Code Systems']];
+  const valueSetDetailsLines = [['Value Set', 'Code System', 'Code', 'Code Description']];
+  for (const de of specs.dataElements.all) {
+    fillElementLines(dataElementLines, profileLines, vsMap, de, specs, config);
+  }
+
+  for (const vs of vsMap.values()) {
+    if (vs) fillValueSetLines(vs, valueSetLines, valueSetDetailsLines);
+  }
+
+  return { profileLines, dataElementLines, valueSetLines, valueSetDetailsLines };
 };
